@@ -1,4 +1,3 @@
-import type { FC } from 'react';
 import React from 'react';
 
 import type { IChildren } from '@/types/common.types';
@@ -7,7 +6,7 @@ type PrimaryHeadingProps = {
   className?: string;
 } & IChildren;
 
-const PrimaryHeading: FC<PrimaryHeadingProps> = ({ children, className }) => {
+const PrimaryHeading = ({ children, className }: PrimaryHeadingProps) => {
   return (
     <h2
       className={`text-[2.4rem] font-bold leading-snug text-gray-900 dark:text-gray-200 md:text-[2.6rem] xl:text-[3.2rem] ${className}`}
